NEW reduction in force policy
This creates a new 0401.15 Reduction in Force policy, built from the preamble of the previous (ie, currently in force) 0401.15.  It also adds the SORD language and specifically recommends a formal RIF declaration.

It also moves the two previous ARs related to RIFs into the new parent policy, and aside from removing the preamble is otherwise unchanged.

fixes #287